Extract cron-related Store methods into impl module.
Follows pattern from 117f63fe44f2d8d3c32691ebdd926f9fbf2d6659 and
multiple prior similar commits. store.rs is now down to < 3K lines.
